# 2x faster (rms) norm cuda kernels (3.7% e2e improvement)