This PR optimized rms_norm and norm cuda kernels for hidden size over 1024, achieving a 2x speedup on kernel and a 3.7% end-to-end generation speedup.

Tested on a V100 GPU with a LLaMA-7B q4_0 model with 4096 hidden size:
```sh
./main -m ./models/7B/ggml-model-q4_0.gguf -p "Building a website can be done in 10 simple steps:" -n 512 -ngl 32 -t 1
```

Before this PR:
```
llama_print_timings:        load time =  1265.11 ms
llama_print_timings:      sample time =   298.76 ms /   512 runs   (    0.58 ms per token,  1713.74 tokens per second)
llama_print_timings: prompt eval time =   290.99 ms /   271 tokens (    1.07 ms per token,   931.30 tokens per second)
llama_print_timings:        eval time =  4628.23 ms /   510 runs   (    9.07 ms per token,   110.19 tokens per second)
llama_print_timings:       total time =  5443.79 ms
```
Single rms_norm kernel time: 8.9us (reported by nsys)

After this PR:
```
llama_print_timings:        load time =  1283.14 ms
llama_print_timings:      sample time =   306.41 ms /   512 runs   (    0.60 ms per token,  1670.99 tokens per second)
llama_print_timings: prompt eval time =   287.84 ms /   271 tokens (    1.06 ms per token,   941.49 tokens per second)
llama_print_timings:        eval time =  4452.48 ms /   510 runs   (    8.73 ms per token,   114.54 tokens per second)
llama_print_timings:       total time =  5268.04 ms
```
Single rms_norm kernel time: 4.5us (reported by nsys)

Kernel speedup: 8.9us/4.5us = 1.98

Per token latency reduced by: 1-(8.73ms/9.07ms) = 3.7%

Similar speedup rate is achieved for norm op tested with a ChatGLM-6B q4_0 model using [chatglm.cpp](https://github.com/li-plus/chatglm.cpp): 8.1 ms/token before this PR v.s. 7.8 ms/token with this PR (3.7% improvement)